@@ -4,8 +4,8 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "dev": "nodemon src/app.ts",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "node src/app.js",
+    "dev": "nodemon src/app.js"
   },
   "keywords": [],
   "author": "",
@@ -25,10 +25,15 @@
     "dotenv": "^16.4.5",
     "express": "^4.19.2",
     "jsonwebtoken": "^9.0.2",
+    "mongoose": "^8.13.2",
     "node-cron": "^3.0.3",
     "nodemon": "^3.1.0",
     "pg": "^8.11.5",
     "ts-node": "^10.9.2",
-    "typescript": "^5.4.5"
-  }
+    "typescript": "^5.8.3"
+  },
+  "devDependencies": {
+    "@types/node-cron": "^3.0.11"
+  },
+  "type": "module"
 }
