@@ -1,26 +1,22 @@
 "use client";
-import BlogCard from "./blogs/blog";
-import { useEffect, useState } from "react";
+import React, { useEffect, useState } from "react";
 import { BlogData } from "@/type";
 import { getBlogsData } from "@/lib/actions/blogs.action";
+import Blog from "@/components/blogs/blog";
 import ClipLoader from "react-spinners/ClipLoader";
 
-const AuthorData = ({
-  searchText,
-  token,
-}: {
-  searchText: string;
-  token: string | undefined;
-}) => {
+interface Props {
+  token: string;
+}
+
+export default function Home({ token }: Props) {
   const [blogsData, setBlogsData] = useState<BlogData[]>([]);
-  const [loading, setLoading] = useState(false);
+  const [loading, setLoading] = useState(true);
 
   useEffect(() => {
     fetchBlogs();
   }, []);
-
   const fetchBlogs = async () => {
-    setLoading(true);
     try {
       const blogs = await getBlogsData(token);
       setBlogsData(blogs);
@@ -31,27 +27,23 @@ const AuthorData = ({
     }
   };
 
-  const filteredBlogs = blogsData.filter((blog) =>
-    blog.title.toLowerCase().includes(searchText.toLowerCase())
-  );
-
   if (loading) {
     return (
       <div className="flex justify-center items-center h-screen">
-        <ClipLoader color="#4B6BFB" size={40} />
+        <ClipLoader color="#4B6BFB" size={40} speedMultiplier={1} />
       </div>
     );
   }
 
   return (
-    <div>
+    <main className="mx-[2rem] lg:mx-[10rem]">
       <div className="text-[#333333] text-[2rem] font-bold py-10">
-        Filtered Blogs
+        Latest Blogs
       </div>
       <div className="grid grid-cols-1 lg:grid-cols-3 gap-6">
-        {filteredBlogs.length > 0 ? (
-          filteredBlogs.map((blog) => (
-            <BlogCard key={blog._id} blog={blog} screentype="authorBlogs" />
+        {blogsData.length > 0 ? (
+          blogsData.map((blog) => (
+            <Blog key={blog._id} blog={blog} screentype="home" />
           ))
         ) : (
           <div className="text-[#333333] text-[1rem] font-medium">
@@ -59,8 +51,6 @@ const AuthorData = ({
           </div>
         )}
       </div>
-    </div>
+    </main>
   );
-};
-
-export default AuthorData;
+}
